+++ akdb/src/opti/rel_eq_comut.c
